Address review feedback and unblock CI
