+++ config/doctrine/CustomerType.mongodb.xml
@@ -18,9 +18,6 @@
         <indexes>
             <index>
                 <key name="value" order="asc" />
-            </index>
-            <index>
-                <key name="value" order="asc" />
                 <key name="ulid" order="asc" />
             </index>
         </indexes>
