+++ tests/Unit/Customer/Application/Resolver/CustomerPatchUpdateResolverTest.php
@@ -6,7 +6,7 @@ namespace App\Tests\Unit\Customer\Application\Resolver;
 
 use App\Core\Customer\Application\DTO\CustomerPatch;
 use App\Core\Customer\Application\Resolver\CustomerPatchUpdateResolver;
-use App\Core\Customer\Application\Resolver\CustomerReferenceResolver;
+use App\Core\Customer\Application\Resolver\CustomerReferenceResolverInterface;
 use App\Core\Customer\Application\Resolver\CustomerUpdateScalarResolver;
 use App\Core\Customer\Application\Transformer\CustomerRelationTransformer;
 use App\Core\Customer\Domain\Entity\Customer;
@@ -18,13 +18,15 @@ use PHPUnit\Framework\MockObject\MockObject;
 
 final class CustomerPatchUpdateResolverTest extends UnitTestCase
 {
-    private CustomerReferenceResolver|MockObject $referenceResolver;
+    private CustomerReferenceResolverInterface|MockObject $referenceResolver;
     private CustomerPatchUpdateResolver $resolver;
 
     protected function setUp(): void
     {
         parent::setUp();
-        $this->referenceResolver = $this->createMock(CustomerReferenceResolver::class);
+        $this->referenceResolver = $this->createMock(
+            CustomerReferenceResolverInterface::class
+        );
         $this->resolver = new CustomerPatchUpdateResolver(
             new CustomerUpdateScalarResolver(),
             new CustomerRelationTransformer($this->referenceResolver)
@@ -77,6 +79,48 @@ final class CustomerPatchUpdateResolverTest extends UnitTestCase
         self::assertSame($type, $update->newType);
         self::assertSame($status, $update->newStatus);
         self::assertTrue($update->newConfirmed);
+    }
+
+    public function testBuildPreservesExistingRelationsWhenReferencesAreOmitted(): void
+    {
+        $dto = new CustomerPatch(
+            id: null,
+            initials: null,
+            email: null,
+            phone: null,
+            leadSource: null,
+            type: null,
+            status: null,
+            confirmed: null,
+        );
+
+        $type = $this->createConfiguredMock(CustomerType::class, [
+            'getUlid' => (string) $this->faker->ulid(),
+        ]);
+        $status = $this->createConfiguredMock(CustomerStatus::class, [
+            'getUlid' => (string) $this->faker->ulid(),
+        ]);
+        $customer = $this->createConfiguredMock(Customer::class, [
+            'getInitials' => 'OLD',
+            'getEmail' => 'old@example.com',
+            'getPhone' => '+999',
+            'getLeadSource' => 'Legacy',
+            'getType' => $type,
+            'getStatus' => $status,
+            'isConfirmed' => true,
+        ]);
+
+        $this->referenceResolver
+            ->expects(self::never())
+            ->method('resolveType');
+        $this->referenceResolver
+            ->expects(self::never())
+            ->method('resolveStatus');
+
+        $update = $this->resolver->build($dto, $customer);
+
+        self::assertSame($type, $update->newType);
+        self::assertSame($status, $update->newStatus);
     }
 
     public function testBuildThrowsWhenTypeIriIsInvalid(): void
